Внесены правки, лабораторная работа 4, вариант 5.
